+++ Sources/RichTextKit/_Deprecated/RichTextDataFormat+Deprecated.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+public extension RichTextDataFormat {
+
+    @available(*, deprecated, renamed: "convertibleFormats")
+    var convertableFormats: [RichTextDataFormat] {
+        convertibleFormats
+    }
+}