+++ RELEASE_NOTES.md
@@ -18,8 +18,10 @@ The library and demo app still updates way to often, and updates too much. For i
 
 * `RichTextDataFormat` has a new `fileFormatText` property.
 * `RichTextDataFormat` has a new `libraryFormats` property that returns all formats except the `vendorArchivedData` format.
-* `RichTextDataFormatMenu` is a new menu for exporting rich text with various formats.
+* `RichTextDataFormatMenu` is a new menu that triggers an action for various formats.
 * `RichTextDataReader` now has ways to get data for the current rich text.
+* `RichTextExportFormatMenu` is a new menu for exporting rich text.
+* `RichTextShareFormatMenu` is a new menu for exporting rich text.
 * `ViewDebug` is a new demo app class that is used for view debugging capabilities.
 * The library is now localized in Danish, German and Norwegian, although some translations probably need adjusting.